requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
pydantic>=2.0.0
typer>=0.12.0
rich>=13.0.0
customtkinter>=5.2.0
pandas>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
